requirements.txt
flask
python-dotenv
flask-MySQLdb